Move more to src/common
