+++ src/content-scripts/components/viewer/report-editor/Report.tsx
@@ -5,7 +5,7 @@ import styles from './ReportEditor.scss';
 import { AnnotationAPIModel } from 'content-scripts/api/annotations';
 import { AnnotationReportAPIModel, DataResponse, Reasons } from 'content-scripts/api/annotation-reports';
 import { PPScopeClass } from 'content-scripts/class_consts';
-import ppGA from 'pp-ga';
+import ppGA from 'common/pp-ga';
 
 interface IReportProps {
   annotation: AnnotationAPIModel;
